requirements.txt
torch
torchvision
pyyaml
wandb
pandas
numpy
pillow
opencv-python
